+++ mission003/ganeodolu/index.html
@@ -97,10 +97,7 @@
                     </div>
                   </div>
                   <button id="add-family-button" class="mdc-icon-button material-icons">add</button>
-
                   <div class="user-family-forms-data"></div>
-
-
                   <footer class="step2-complete-btn mdc-dialog__actions">
                     <button type="submit" class="mdc-button mdc-button--raised dialog__button">
                       <span class="mdc-button__ripple"></span>제출
@@ -119,8 +116,6 @@
                   </button>
                 </footer>
               </div>
-              <!-- <div id="step4" class="hidden" > -->
-                
             </div>
           </form>
         </div>
